% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reproduce.R
\name{randomMatches}
\alias{randomMatches}
\title{Random Mating}
\usage{
randomMatches(individuals, number_of_couples)
}
\arguments{
\item{individuals}{Data frame containing id, sex, warner-status,
mother id and father id.}

\item{number_of_couples}{Number of couples to form in the current
generation.}
}
\description{
Default function to govern mating.  Each fertile female mates
with a randomly-selected male.
}
