% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simaltruist-package.R
\docType{package}
\name{simaltruist}
\alias{simaltruist}
\alias{simaltruist-package}
\title{Altruistic Behavior in Animal Populations}
\description{
For use by students and colleagues in evolutionary biology.
}
