% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulate.R
\name{simulate}
\alias{simulate}
\title{Simulate Spread of Gene for Altruism}
\usage{
simulate(initial_pop = list(m0 = 90, m1 = 0, m2 = 10, f0 = 90, f1 = 0, f2 =
  10), average_litter_size = 5, birth_rate_natural = 0.05,
  death_rate_natural = 0, prob_attack = 0.2, warner_death_prob = 0.4,
  nonwarner_death_prob = 0.2, hider_death_prob = 0, sim_gens = 2,
  capacity = 2000, mating_behavior = NULL, attack_behavior = NULL,
  graph = TRUE)
}
\arguments{
\item{initial_pop}{List comprised of six named elements:
\describe{
  \item{m0}{initial number of males with 0 altruist alleles;}
  \item{m1}{initial number of males with 1 altruist allele;}
  \item{m2}{initial number of males with 2 altruist alleles;}
  \item{f0}{initial number of females with 0 altruist alleles;}
  \item{f1}{initial number of females with 1 altruist allele;}
  \item{f2}{initial number of females with 2 altruist alleles.}
}}

\item{average_litter_size}{Mean size of a litter.}

\item{birth_rate_natural}{Birth rate for the population.}

\item{death_rate_natural}{Death rate for a size-zero population. Rises
linearly to \code{birth_rate_natrual} as populaiion approaches
\code{capacity}.}

\item{prob_attack}{The probability of a predator attack in any given
generation.}

\item{warner_death_prob}{Probability that an individual who warns
others during an attack will be killed.}

\item{nonwarner_death_prob}{Probability of death for an individual who
does not warn others but who was not forewarned by a warner.}

\item{hider_death_prob}{Probability of death during an attack for an
individual who accepts a warning.}

\item{sim_gens}{Number of generations to simulate.}

\item{capacity}{Carrying capacity of the population.}

\item{mating_behavior}{Custom function to govern how eligible
partners form couples.}

\item{attack_behavior}{Custom function to govern behavior of
warners when population is under attack.}

\item{graph}{If \code{TRUE}, provides a graph of the total
per-capita warner alleles in population over the generations.}
}
\value{
A data frame with information on the population at
each generation.  Each row describes the population at the end of
a single birth-death-attack cycle.   Variables are:
\describe{
  \item{populationSize}{total population}
  \item{males}{total number of males}
  \item{males0}{number of males with no alleles for altruism}
  \item{males1}{number of males with one allele for altruism}
  \item{males2}{number of males with two alleles for altruism}
  \item{females}{total number of females}
  \item{females0}{number of females with no alleles for altruism}
  \item{females1}{number of females with one allele for altruism}
  \item{females2}{number of females with two alleles for altruism}
}
}
\description{
Simulate Spread of Gene for Altruism
}
\note{
For details on \code{mating_behavior} and \code{attack_behavior}
consult the
\href{https://homerhanumat.github.io/simaltruist}{package documentation}.
}
\examples{
\dontrun{
# use defaults, get a graph:
pop <- simulate(sim_gens = 400)
# attacks are infrequent, and it's dangerous to warn:
pop <- simulate(sim_gens = 200,
                warner_death_prob = 0.8,
                attack_prob = 0.05)
# use an alternative mating function exported by package:
pop <- simulate(sim_gens = 200,
                warner_death_prob = 0.8,
                mating_behavior = list(
                  fn = sexualSelection,
                  args = list(
                    matrix(
                      c(1, 5, 10, 1, 1, 1, 10, 5, 1),
                      nrow = 3,
                      ncol = 3))))
}
}
