% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/attack.R
\name{warnRelativesGraph}
\alias{warnRelativesGraph}
\title{Warning Relatives When Under Attack (Graph Version)}
\usage{
warnRelativesGraph(individuals, number_warned, warner_death_prob,
  nonwarner_death_prob, hider_death_prob, order = 2, relGraph,
  dominant = FALSE)
}
\arguments{
\item{individuals}{Data frame containing id, sex, warner-status,
mother id and father id.  Value provided.}

\item{number_warned}{Maximum number of individuals to warn.
Value provided.}

\item{warner_death_prob}{As in \code{simulate}. Value provided.}

\item{nonwarner_death_prob}{As in \code{simulate}. Value provided.}

\item{hider_death_prob}{As in \code{simulate}. Value provided.}

\item{order}{Maximum distance a relative can have and still be warned.}

\item{relGraph}{The current relationship graph.  Value provided.}

\item{dominant}{If \code{TRUE}, individuals with 1 or more altruistic
alleles can warn.  Default version sets this to \code{FALSE}.}
}
\value{
A numerical vector of death probabilities in which the
\eqn{i^{\text{th}}} element corresponds to row \eqn{i} of
\code{individuals}.
}
\description{
Default function to govern behavior of warners under attack.
Warner warns a specified number of the most closely-related
individuals.  May be used as an alternative function in which
the values of \code{warnable_relationship} and \code{dominant}
are specified.
}
\examples{
\dontrun{
pop <- simulate(sim_gens = 200,
                relationship_method = "graph",
                attack_behavior = list(
                  fn = warnRelativesGraph,
                  args = list(
                    number_warned = 10,
                    dominant = TRUE)))
}
}
