% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/attack.R
\name{warnWarners}
\alias{warnWarners}
\title{Warning Other Warners}
\usage{
warnWarners(individuals, number_warned, warner_death_prob, nonwarner_death_prob,
  hider_death_prob, dominant)
}
\arguments{
\item{individuals}{Data frame containing id, sex, warner-status,
mother id and father id.}

\item{number_warned}{Maximum number of individuals to warn.}

\item{warner_death_prob}{As in \code{simulate}.}

\item{nonwarner_death_prob}{As in \code{simulate}.}

\item{hider_death_prob}{As in \code{simulate}.}

\item{dominant}{If \code{TRUE}, individuals with 1 or more altruistic
alleles can warn.}
}
\description{
Example of an lternative function to govern behavior of warners
under attack.
Warner warns a specified number of the most closely-related
individuals.  Warners only warn other warners.
}
