% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulate.R
\name{simulate}
\alias{simulate}
\title{Simulate Spread of Gene for Altruism}
\usage{
simulate(average_litter_size = 5, initial_males = 100,
  initial_alt_males = 10, initial_females = 100, initial_alt_females = 10,
  birth_rate_natural = 0.05, death_rate_natural = 0, prob_attack = 0.2,
  warner_death_prob = 0.4, nonwarner_death_prob = 0.2,
  hider_death_prob = 0, sim_gens = 2, capacity = 2000,
  mating_behavior = NULL, attack_behavior = NULL, graph = TRUE)
}
\arguments{
\item{average_litter_size}{Mean size of a litter.}

\item{initial_males}{Number of males in initial population.}

\item{initial_alt_males}{Number of males in initial population with
two alleles for altruism (all other initial males have 0 such alleles).}

\item{initial_females}{Number of females in initial population.}

\item{initial_alt_females}{Number of females in initial population with
two alleles for altruism (all other initial females have 0 such alleles).}

\item{birth_rate_natural}{Birth rate for the population.}

\item{death_rate_natural}{Death rate for a size-zero population. Rises
linearly to \code{birth_rate_natrual} as populaiion approaches
\code{capacity}.}

\item{prob_attack}{The probability of a predator attack in any given
generation.}

\item{warner_death_prob}{Probability that an individual who warns
others during an attack will be killed.}

\item{nonwarner_death_prob}{Probability of death for an individual who
does not warn others but who was not forewarned by a warner.}

\item{hider_death_prob}{Probability of death during an attack for an
individual who accepts a warning.}

\item{sim_gens}{Number of generations to simulate.}

\item{capacity}{Carrying capacity of the population.}

\item{mating_behavior}{Custom function to govern how eligible
partners form couples.}

\item{attack_behavior}{Custom function to govern behavior of
warners when population is under attack.}

\item{graph}{If \code{TRUE}, provides a graph of the total
per-capita warner alleles in population over the generations.}
}
\value{
A data frame with information on the population at
each generation.
}
\description{
Simulate Spread of Gene for Altruism
}
\examples{
\dontrun{
# use defaults, get a graph:
pop <- simulate(sim_gens = 400)
# attacks are infrequent, and it's dangerous to warn:
pop <- simulate(sim_gens = 200,
                warner_death_prob = 0.8,
                attack_prob = 0.05)
# use an alternative mating function exported by package:
pop <- simulate(sim_gens = 200,
                warner_death_prob = 0.8,
                mating_behavior = list(
                  fn = sexualSelection,
                  args = list(
                    matrix(
                      c(1, 5, 10, 1, 1, 1, 10, 5, 1),
                      nrow = 3,
                      ncol = 3))))
}
}
