% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/attack.R
\name{warnWarners}
\alias{warnWarners}
\title{Warning Other Warners}
\usage{
warnWarners(individuals, number_warned, warner_death_prob, nonwarner_death_prob,
  hider_death_prob, dominant)
}
\arguments{
\item{individuals}{Data frame containing id, sex, warner-status,
mother id and father id.Value provided.}

\item{number_warned}{Maximum number of individuals to warn.}

\item{warner_death_prob}{As in \code{simulate}.  Value provided.}

\item{nonwarner_death_prob}{As in \code{simulate}.  Value provided.}

\item{hider_death_prob}{As in \code{simulate}.  Value provided.}

\item{dominant}{If \code{TRUE}, individuals with 1 or more altruistic
alleles can warn.}
}
\value{
A numerical vector of death probabilities in which the
\eqn{i^{\text{th}}} element corresponds to row \eqn{i} of
\code{individuals}.
}
\description{
Example of an alternative function to govern behavior of warners
under attack.  User must set values of \code{number_warned} and
\code{dominant}.
Warner warns a specified number of the most closely-related
individuals.  Warners only warn other warners.
}
\examples{
\dontrun{
pop <- simulate(sim_gens = 200,
                attack_behavior = list(
                  fn = warnWarners,
                  args = list(
                    number_warned = 7,
                    dominant = TRUE)))
}
}
